Refactor navigation reducer responsibilities
